human-bot-mvp/bot-engine/requirements.txt
# Core Dependencies
playwright>=1.40.0
asyncio>=3.4.3

# API Framework
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
websockets>=12.0
pydantic>=2.5.0

# Database
sqlalchemy>=2.0.23
aiosqlite>=0.19.0

# Configuration
pyyaml>=6.0.1
python-dotenv>=1.0.0

# HTTP Client
httpx>=0.25.0

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1

# Future AI Integration
# transformers>=4.35.0
# torch>=2.1.0
# scikit-learn>=1.3.2
